+++ DSCResources/MSFT_xSQLServerScript/MSFT_xSQLServerScript.psm1
@@ -1,10 +1,39 @@
-﻿$currentPath = Split-Path -Parent $MyInvocation.MyCommand.Path
-Write-Verbose -Message "CurrentPath: $currentPath"
+﻿$script:currentPath = Split-Path -Path $MyInvocation.MyCommand.Path -Parent
+Import-Module -Name (Join-Path -Path (Split-Path -Path (Split-Path -Path $script:currentPath -Parent) -Parent) -ChildPath 'xSQLServerHelper.psm1')
 
-# Load Common Code
-Import-Module $currentPath\..\..\xSQLServerHelper.psm1 -Verbose:$false -ErrorAction Stop
+<#
+    .SYNOPSIS
+        Returns the current state of the SQL Server features.
 
+    .PARAMETER ServerInstance
+        The name of an instance of the Database Engine. For a default instance, only specify the computer name. For a named instances,
+        use the format ComputerName\InstanceName.
 
+    .PARAMETER SetFilePath
+        Path to the T-SQL file that will perform Set action.
+
+    .PARAMETER GetFilePath
+        Path to the T-SQL file that will perform Get action.
+        Any values returned by the T-SQL queries will also be returned by the cmdlet Get-DscConfiguration through the `GetResult` property.
+
+    .PARAMETER TestFilePath
+        Path to the T-SQL file that will perform Test action.
+        Any script that does not throw an error or returns null is evaluated to true.
+        The cmdlet Invoke-SqlCmd treats T-SQL Print statements as verbose text, and will not cause the test to return false.
+
+    .PARAMETER Credential
+        The credentials to authenticate with, using SQL Authentication. To authenticate using Windows Authentication, assign the credentials
+        to the built-in parameter `PsDscRunAsCredential`. If both parameters `Credential` and `PsDscRunAsCredential` are not assigned,
+        then SYSTEM account will be used to authenticate using Windows Authentication.
+
+    .PARAMETER Variable
+        Specifies, as a string array, a sqlcmd scripting variable for use in the sqlcmd script, and sets a value for the variable.
+        Use a Windows PowerShell array to specify multiple variables and their values. For more information how to use this,
+        please go to the help documentation for [Invoke-Sqlcmd](https://technet.microsoft.com/en-us/library/mt683370.aspx).
+
+    .OUTPUTS
+        Hash table containing key 'GetResult' which holds the value of the result from the SQL script that was ran from the parameter 'GetFilePath'.
+#>
 function Get-TargetResource
 {
     [CmdletBinding()]
@@ -28,23 +57,24 @@ function Get-TargetResource
         $TestFilePath,
 
         [System.Management.Automation.PSCredential]
+        [System.Management.Automation.Credential()]
         $Credential,
 
         [System.String[]]
         $Variable
-    )   
+    )
 
     $result = Invoke-SqlScript -ServerInstance $ServerInstance -SqlScriptPath $GetFilePath `
                 -Credential $Credential -Variable $Variable -ErrorAction Stop
 
     $getResult = Out-String -InputObject $result
-        
+
     $returnValue = @{
         ServerInstance = [System.String] $ServerInstance
         SetFilePath = [System.String] $SetFilePath
         GetFilePath = [System.String] $GetFilePath
         TestFilePath = [System.String] $TestFilePath
-        Username = [System.Management.Automation.PSCredential] $Credential
+        Username = [System.Object] $Credential
         Variable = [System.String[]] $Variable
         GetResult = [System.String[]] $getresult
     }
@@ -52,6 +82,36 @@ function Get-TargetResource
     $returnValue
 }
 
+<#
+    .SYNOPSIS
+        Returns the current state of the SQL Server features.
+
+    .PARAMETER ServerInstance
+        The name of an instance of the Database Engine. For a default instance, only specify the computer name. For a named instances,
+        use the format ComputerName\InstanceName.
+
+    .PARAMETER SetFilePath
+        Path to the T-SQL file that will perform Set action.
+
+    .PARAMETER GetFilePath
+        Path to the T-SQL file that will perform Get action.
+        Any values returned by the T-SQL queries will also be returned by the cmdlet Get-DscConfiguration through the `GetResult` property.
+
+    .PARAMETER TestFilePath
+        Path to the T-SQL file that will perform Test action.
+        Any script that does not throw an error or returns null is evaluated to true.
+        The cmdlet Invoke-SqlCmd treats T-SQL Print statements as verbose text, and will not cause the test to return false.
+
+    .PARAMETER Credential
+        The credentials to authenticate with, using SQL Authentication. To authenticate using Windows Authentication, assign the credentials
+        to the built-in parameter `PsDscRunAsCredential`. If both parameters `Credential` and `PsDscRunAsCredential` are not assigned,
+        then SYSTEM account will be used to authenticate using Windows Authentication.
+
+    .PARAMETER Variable
+        Specifies, as a string array, a sqlcmd scripting variable for use in the sqlcmd script, and sets a value for the variable.
+        Use a Windows PowerShell array to specify multiple variables and their values. For more information how to use this,
+        please go to the help documentation for [Invoke-Sqlcmd](https://technet.microsoft.com/en-us/library/mt683370.aspx).
+#>
 function Set-TargetResource
 {
     [CmdletBinding()]
@@ -74,6 +134,7 @@ function Set-TargetResource
         $TestFilePath,
 
         [System.Management.Automation.PSCredential]
+        [System.Management.Automation.Credential()]
         $Credential,
 
         [System.String[]]
@@ -84,7 +145,37 @@ function Set-TargetResource
                 -Credential $Credential -Variable $Variable -ErrorAction Stop
 }
 
+<#
+    .SYNOPSIS
+        Returns the current state of the SQL Server features.
 
+    .PARAMETER ServerInstance
+        The name of an instance of the Database Engine. For a default instance, only specify the computer name. For a named instances,
+        use the format ComputerName\InstanceName.
+
+    .PARAMETER SetFilePath
+        Path to the T-SQL file that will perform Set action.
+
+    .PARAMETER GetFilePath
+        Path to the T-SQL file that will perform Get action.
+        Any values returned by the T-SQL queries will also be returned by the cmdlet Get-DscConfiguration through the `GetResult` property.
+
+    .PARAMETER TestFilePath
+        Path to the T-SQL file that will perform Test action.
+        Any script that does not throw an error or returns null is evaluated to true.
+        The cmdlet Invoke-SqlCmd treats T-SQL Print statements as verbose text, and will not cause the test to return false.
+
+    .PARAMETER Credential
+        The credentials to authenticate with, using SQL Authentication. To authenticate using Windows Authentication, assign the credentials
+        to the built-in parameter `PsDscRunAsCredential`. If both parameters `Credential` and `PsDscRunAsCredential` are not assigned,
+        then SYSTEM account will be used to authenticate using Windows Authentication.
+
+    .PARAMETER Variable
+        Specifies, as a string array, a sqlcmd scripting variable for use in the sqlcmd script, and sets a value for the variable.
+        Use a Windows PowerShell array to specify multiple variables and their values. For more information how to use this,
+        please go to the help documentation for [Invoke-Sqlcmd](https://technet.microsoft.com/en-us/library/mt683370.aspx).
+
+#>
 function Test-TargetResource
 {
     [CmdletBinding()]
@@ -108,6 +199,7 @@ function Test-TargetResource
         $TestFilePath,
 
         [System.Management.Automation.PSCredential]
+        [System.Management.Automation.Credential()]
         $Credential,
 
         [System.String[]]
@@ -115,11 +207,11 @@ function Test-TargetResource
     )
 
     try
-    {   
+    {
         $result = Invoke-SqlScript -ServerInstance $ServerInstance -SqlScriptPath $TestFilePath `
                 -Credential $Credential -Variable $Variable -ErrorAction Stop
 
-        if($result -eq $null)
+        if($null -eq $result)
         {
             return $true
         }
@@ -135,6 +227,25 @@ function Test-TargetResource
     }
 }
 
+<#
+    .SYNOPSIS
+        Execute an SQL script located in a file on disk.
+
+    .PARAMETER ServerInstance
+        The name of an instance of the Database Engine.
+        For default instances, only specify the computer name. For named instances, use the format ComputerName\InstanceName.
+
+    .PARAMETER SqlScriptPath
+        Path to SQL script file that will be executed.
+
+    .PARAMETER Credential
+        The credentials to use to authenticate using SQL Authentication. To authenticate using Windows Authentication, assing the credentials
+        to the built-in parameter 'PsDscRunAsCredential'. If both parameters 'Credential' and 'PsDscRunAsCredential' are not assigned, then
+        the SYSTEM account will be used to authenticate using Windows Authentication.
+
+    .PARAMETER Variable
+        Creates a sqlcmd scripting variable for use in the sqlcmd script, and sets a value for the variable.
+#>
 function Invoke-SqlScript
 {
     param
@@ -148,6 +259,7 @@ function Invoke-SqlScript
         $SqlScriptPath,
 
         [System.Management.Automation.PSCredential]
+        [System.Management.Automation.Credential()]
         $Credential,
 
         [System.String[]]
@@ -159,7 +271,7 @@ function Invoke-SqlScript
     if($null -ne $Credential)
     {
         $null = $PSBoundParameters.Add("Username", $Credential.UserName)
-        $null = $PSBoundParameters.Add("Password", $Credential.GetNetworkCredential().password)   
+        $null = $PSBoundParameters.Add("Password", $Credential.GetNetworkCredential().password)
     }
 
     $null = $PSBoundParameters.Remove("Credential")
@@ -169,4 +281,3 @@ function Invoke-SqlScript
 }
 
 Export-ModuleMember -Function *-TargetResource
-